small fixes in objects

+++ space_object.py
@@ -5,7 +5,7 @@ import numpy as np
 class SpaceObject:
     def __init__(self, position: np.array = None, mass: float = 1.0):
         if position is None:
-            self._position = np.array([0, 0])
+            self._position = np.array([0, 0]).astype(float)
         else:
             self._position = position
